Move interfaces to their own files
