+++ src/core/i-minimal-chart.ts
@@ -1,0 +1,7 @@
+export interface IMinimalChart {
+    render(): void;
+    redraw(): void;
+    filterAll(): void;
+    focus?(): void;
+    dispose?(): void;
+}
