+++ src/core/index.ts
@@ -2,7 +2,6 @@ export * from './filters/index';
 
 export * from './bad-argument-exception';
 export * from './chart-group';
-export * from './chart-group-types';
 export * from './chart-registry';
 export * from './config';
 export * from './constants';
@@ -11,6 +10,9 @@ export * from './d3compat';
 export * from './events';
 export * from './filter-storage';
 export * from './invalid-state-exception';
+export * from './i-filter-storage';
+export * from './i-chart-group';
+export * from './i-minimal-chart';
 export * from './logger';
 export * from './printers';
 export * from './types';
