+++ src/core/types.ts
@@ -1,6 +1,6 @@
 import { Selection } from 'd3-selection';
 import { BaseMixin } from '../base/base-mixin';
-import { IChartGroup } from './chart-group-types';
+import { IChartGroup } from './i-chart-group';
 
 export type ChartParentType = string | BaseMixin | Selection<Element, undefined, Element, unknown>;
